виправив всі помилки
