+++ src/main/java/core/basesyntax/Model/Cat.java
@@ -1,4 +1,4 @@
-package core.basesyntax.Model;
+package core.basesyntax.model;
 
 public class Cat {
     private int age;
@@ -7,10 +7,6 @@ public class Cat {
 
     @Override
     public String toString() {
-        return "Cat{" +
-                "age=" + age +
-                ", name='" + name + '\'' +
-                ", ownerName='" + ownerName + '\'' +
-                '}';
+        return "Cat" + "age=" + age + ", name=" + name + ", ownerName=" + ownerName;
     }
 }
