+++ src/main/java/core/basesyntax/Model/Dog.java
@@ -1,4 +1,4 @@
-package core.basesyntax.Model;
+package core.basesyntax.model;
 
 public class Dog {
 }
